# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life RPG API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Character {
  id: number;
  name: string;
  class: string;
  overallLevel: number;
  totalXp: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpdateCharacterBody {
  name?: string;
  class?: string;
}

export interface Stat {
  id: number;
  name: string;
  displayName: string;
  xp: number;
  level: number;
  title: string;
  createdAt: string;
  updatedAt: string;
}

export interface AddXpBody {
  amount: number;
  reason?: string;
}

export interface XpGain {
  statName: string;
  amount: number;
}

export interface Activity {
  id: number;
  name: string;
  description: string;
  xpGains: XpGain[];
  xpLosses: XpGain[];
  category: string;
}

export interface XpChange {
  statName: string;
  amount: number;
  reason: string;
}

export interface LevelUp {
  statName: string;
  newLevel: number;
  newTitle: string;
}

export interface DailyLog {
  id: number;
  date: string;
  activities: string[];
  totalXpGained: number;
  totalXpLost: number;
  xpChanges: XpChange[];
  newLevelUps: LevelUp[];
  streaksUpdated: string[];
  rewardsEarned: string[];
  punishmentsAssigned: string[];
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface SubmitDailyLogBody {
  /** List of activity names done today */
  activities: string[];
  /** @nullable */
  notes?: string | null;
  gymDone?: boolean;
  runningDone?: boolean;
  basketballDone?: boolean;
  studyDone?: boolean;
  deepWorkDone?: boolean;
  pianoDone?: boolean;
  sleepHours?: number;
  ateJunkFood?: boolean;
  phoneHours?: number;
  socializedToday?: boolean;
  plannedDay?: boolean;
  coldShower?: boolean;
  meditatedToday?: boolean;
  drankWater?: boolean;
}

export interface Achievement {
  id: number;
  name: string;
  description: string;
  icon: string;
  /** @nullable */
  unlockedAt?: string | null;
  unlocked: boolean;
  requirement: string;
  xpBonus: number;
  createdAt: string;
}

export interface Reward {
  id: number;
  type: string;
  name: string;
  description: string;
  used: boolean;
  /** @nullable */
  usedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface Punishment {
  id: number;
  type: string;
  description: string;
  xpPenalty: number;
  /** @nullable */
  deadline?: string | null;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface DailyLogResult {
  log: DailyLog;
  xpChanges: XpChange[];
  levelUps: LevelUp[];
  newAchievements: Achievement[];
  rewardsEarned: Reward[];
  punishmentsAssigned: Punishment[];
}

export interface DailyQuest {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  statReward: string;
  completed: boolean;
  date: string;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface SideQuest {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  statReward: string;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface CreateSideQuestBody {
  title: string;
  description: string;
  xpReward: number;
  statReward: string;
}

export interface MainQuest {
  id: number;
  title: string;
  description: string;
  xpReward: number;
  completed: boolean;
  /** @nullable */
  completedAt?: string | null;
  createdAt: string;
}

export interface CreateMainQuestBody {
  title: string;
  description: string;
  xpReward: number;
}

export interface Streak {
  id: number;
  name: string;
  displayName: string;
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  lastActivityDate?: string | null;
  updatedAt: string;
}

export interface BossFight {
  id: number;
  name: string;
  description: string;
  eventType: string;
  xpReward: number;
  statsInvolved: string[];
  completed: boolean;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  scheduledFor?: string | null;
  createdAt: string;
}

export interface CreateBossFightBody {
  name: string;
  description: string;
  eventType: string;
  xpReward: number;
  statsInvolved: string[];
  /** @nullable */
  scheduledFor?: string | null;
}

export type CompleteBossFightBodyResult =
  (typeof CompleteBossFightBodyResult)[keyof typeof CompleteBossFightBodyResult];

export const CompleteBossFightBodyResult = {
  victory: "victory",
  defeat: "defeat",
  draw: "draw",
} as const;

export interface CompleteBossFightBody {
  result: CompleteBossFightBodyResult;
  xpEarned: number;
}

export type GetDailyLogsParams = {
  limit?: number;
};