# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Life RPG API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /daily-log:
    get:
      operationId: getDailyLogs
      tags: [dailyLog]
      summary: Get daily logs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of daily logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyLog"
    post:
      operationId: submitDailyLog
      tags: [dailyLog]
      summary: Submit a daily log for today
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitDailyLogBody"
      responses:
        "201":
          description: Daily log result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLogResult"
  /daily-log/today:
    get:
      operationId: getTodayLog
      tags: [dailyLog]
      summary: Get today's log if it exists
      responses:
        "200":
          description: Today's log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyLog"
        "404":
          description: No log for today
...
  /punishments/{id}/complete:
    post:
      operationId: completePunishment
      tags: [punishments]
      summary: Mark punishment as completed
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Completed punishment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Punishment"
...
  /boss-fights:
    get:
      operationId: getBossFights
      tags: [bossFights]
      summary: Get boss fights
      responses:
        "200":
          description: Boss fights
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BossFight"
    post:
      operationId: createBossFight
      tags: [bossFights]
      summary: Create a boss fight event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBossFightBody"
      responses:
        "201":
          description: Created boss fight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossFight"
  /boss-fights/{id}/complete:
    post:
      operationId: completeBossFight
      tags: [bossFights]
      summary: Complete a boss fight and award XP
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteBossFightBody"
      responses:
        "200":
          description: Completed boss fight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BossFight"
components:
  schemas:
...
    DailyLog:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
          format: date-time
        activities:
          type: array
          items:
            type: string
        totalXpGained:
          type: integer
        totalXpLost:
          type: integer
        xpChanges:
          type: array
          items:
            $ref: "#/components/schemas/XpChange"
        newLevelUps:
          type: array
          items:
            $ref: "#/components/schemas/LevelUp"
        streaksUpdated:
          type: array
          items:
            type: string
        rewardsEarned:
          type: array
          items:
            type: string
        punishmentsAssigned:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, date, activities, totalXpGained, totalXpLost, xpChanges, newLevelUps, streaksUpdated, rewardsEarned, punishmentsAssigned, createdAt]
    XpChange:
      type: object
      properties:
        statName:
          type: string
        amount:
          type: integer
        reason:
          type: string
      required: [statName, amount, reason]
    LevelUp:
      type: object
      properties:
        statName:
          type: string
        newLevel:
          type: integer
        newTitle:
          type: string
      required: [statName, newLevel, newTitle]
    SubmitDailyLogBody:
      type: object
      properties:
        activities:
          type: array
          items:
            type: string
          description: List of activity names done today
        notes:
          type: ["string", "null"]
        gymDone:
          type: boolean
        runningDone:
          type: boolean
        basketballDone:
          type: boolean
        studyDone:
          type: boolean
        deepWorkDone:
          type: boolean
        pianoDone:
          type: boolean
        sleepHours:
          type: number
        ateJunkFood:
          type: boolean
        phoneHours:
          type: number
        socializedToday:
          type: boolean
        plannedDay:
          type: boolean
        coldShower:
          type: boolean
        meditatedToday:
          type: boolean
        drankWater:
          type: boolean
      required: [activities]
    DailyLogResult:
      type: object
      properties:
        log:
          $ref: "#/components/schemas/DailyLog"
        xpChanges:
          type: array
          items:
            $ref: "#/components/schemas/XpChange"
        levelUps:
          type: array
          items:
            $ref: "#/components/schemas/LevelUp"
        newAchievements:
          type: array
          items:
            $ref: "#/components/schemas/Achievement"
        rewardsEarned:
          type: array
          items:
            $ref: "#/components/schemas/Reward"
        punishmentsAssigned:
          type: array
          items:
            $ref: "#/components/schemas/Punishment"
      required: [log, xpChanges, levelUps, newAchievements, rewardsEarned, punishmentsAssigned]
...
    Reward:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        name:
          type: string
        description:
          type: string
        used:
          type: boolean
        usedAt:
          type: ["string", "null"]
          format: date-time
        expiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, type, name, description, used, createdAt]
    Punishment:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        description:
          type: string
        xpPenalty:
          type: integer
        deadline:
          type: ["string", "null"]
          format: date-time
        completed:
          type: boolean
        completedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, type, description, xpPenalty, completed, createdAt]
    Achievement:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        icon:
          type: string
        unlockedAt:
          type: ["string", "null"]
          format: date-time
        unlocked:
          type: boolean
        requirement:
          type: string
        xpBonus:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, icon, unlocked, requirement, xpBonus, createdAt]
    BossFight:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        eventType:
          type: string
        xpReward:
          type: integer
        statsInvolved:
          type: array
          items:
            type: string
        completed:
          type: boolean
        result:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
          format: date-time
        scheduledFor:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, description, eventType, xpReward, statsInvolved, completed, createdAt]
    CreateBossFightBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        eventType:
          type: string
        xpReward:
          type: integer
        statsInvolved:
          type: array
          items:
            type: string
        scheduledFor:
          type: ["string", "null"]
          format: date-time
      required: [name, description, eventType, xpReward, statsInvolved]
    CompleteBossFightBody:
      type: object
      properties:
        result:
          type: string
          enum: [victory, defeat, draw]
        xpEarned:
          type: integer
      required: [result, xpEarned]