# lib/api-zod/src/generated/types/bossFight.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life RPG API
 * OpenAPI spec version: 0.1.0
 */

export interface BossFight {
  id: number;
  name: string;
  description: string;
  eventType: string;
  xpReward: number;
  statsInvolved: string[];
  completed: boolean;
  /** @nullable */
  result?: string | null;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  scheduledFor?: Date | null;
  createdAt: Date;
}