# lib/api-zod/src/generated/types/completeBossFightBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { CompleteBossFightBodyResult } from "./completeBossFightBodyResult";

export interface CompleteBossFightBody {
  result: CompleteBossFightBodyResult;
  xpEarned: number;
}