# lib/api-zod/src/generated/types/dailyLogResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Life RPG API
 * OpenAPI spec version: 0.1.0
 */
import type { Achievement } from "./achievement";
import type { DailyLog } from "./dailyLog";
import type { LevelUp } from "./levelUp";
import type { Punishment } from "./punishment";
import type { Reward } from "./reward";
import type { XpChange } from "./xpChange";

export interface DailyLogResult {
  log: DailyLog;
  xpChanges: XpChange[];
  levelUps: LevelUp[];
  newAchievements: Achievement[];
  rewardsEarned: Reward[];
  punishmentsAssigned: Punishment[];
}